Correct a regression introduced in 78bbf63c76bf.
If the local hostname is routable, it will appears in
utils.localHostnames. However, in divideHosts we only
want to compare the hostname to the ones that are not
routable i.e. : 127.0.0.1, localhost and ::1. Therefore
we need to only compare to those, and not the one in
utils.localHostnames. Also added IPV6 localhost for
bonus points.